Time: 318 ms (57.94%), Space: 90.4 MB (44.71%) - LeetHub

+++ 1856-maximum-subarray-min-product/1856-maximum-subarray-min-product.cpp
@@ -50,7 +50,7 @@ public:
         {
             int l = left[i];
             int r = right[i];
-            long summ = prefixSum[r] - (l>0? prefixSum[l-1]:0);
+            long long summ = prefixSum[r] - (l>0? prefixSum[l-1]:0);
             
             temp = (nums[i]*(summ));
             maxi = max(maxi, temp);
